requirements.txt
python-dotenv==1.0.0
pandas==2.2.3
numpy>=1.26.0
requests==2.31.0
jsonrpclib-pelix==0.4.3
python-dateutil==2.8.2
